@@ -22,12 +22,12 @@
 		}
 	},
 	"require": {
-		"php": ">=5.6",
+		"php": ">=8.0",
 		"ext-gd": "*"
 	},
 	"require-dev": {
-		"phpunit/phpunit": ">=5.7",
-		"symfony/var-dumper": ">=3.4",
+		"phpunit/phpunit": ">=9.5",
+		"symfony/var-dumper": ">=5.4",
 		"twig/twig": "<3.0"
 	},
 	"suggest": {
@@ -35,7 +35,7 @@
 	},
 	"extra": {
 		"branch-alias": {
-			"dev-master": "2.x-dev"
+			"dev-master": "3.x-dev"
 		}
 	}
 }
